Add Basic MacOS Support
[Problem]
Attempting to build on a MacOS device currently fails
due to a variety of issues. This includes:

- Referencing "malloc.h"
- Missing checks for Apple compiler flags
- Incompatible compiler and linker flags for MacOS' clang gcc version
- The ARRAY_SIZE macro not being compatible with the MacOS' clang gcc version
- Incompatible sound library (asound)

[Solution]
This change fixes the above issues and updates the documentation
for how to use on MacOS. Sound currently is a work in progress
as a new sound library is searched for to use.

[Testing]
Built and confirmed that the simulator runs in XQuartz now, ensuring
that a few different games are up and running.