Fix Chowa Name Race Condition
[Problem]
When multiple Chowa hatch in the same frame, only the first one to hatch is able to be named since the others won't trigger the state changes.

[Solution]
Ensure we stop processing Chowa states until after the names have been entered.

[Testing]
Confirmed working with 2 and 3 Chowa hatching at the same time.