Update introMode.c
Adjust routing at the end of the tutorial to determine if the user is dumped into swsnCreator or Main Menu.
Adjusted tutorial language a little bit if this is the second time they do the tutorial.
This is to prevent a new random swsn generation when tossed over to creator mode, and to give clearer instructions on how to edit the swsn.